+++ src/plugins/extra/widgets/index.ts
@@ -1,2 +1,4 @@
 export { default as Search } from './Search';
 export { default as SearchSettings } from './SearchSettings';
+export { default as Weather } from './Weather';
+export { default as WeatherSettings } from './WeatherSettings';