+++ CHANGELOG.md
@@ -5,7 +5,7 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/).
 
 ## [Unreleased]
 ### Added
-- New widget (core): search box.
+- New widgets (core): search box and weather.
 - Open newtab page in Chrome on extension install and browser action button clicked.
 - Optional email field for the feedback form.